refactor: #161 <Booking> service 디렉토리 생성 및 이동
