+++ com.taken_seat.booking-service/src/main/java/com/taken_seat/booking_service/booking/application/service/BookingClientService.java
@@ -1,4 +1,4 @@
-package com.taken_seat.booking_service.booking.application;
+package com.taken_seat.booking_service.booking.application.service;
 
 import java.util.UUID;
