+++ com.taken_seat.booking-service/src/test/java/com/taken_seat/booking_service/booking/infrastructure/service/ConcurrencyTest.java
@@ -16,8 +16,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.taken_seat.booking_service.booking.application.BookingService;
 import com.taken_seat.booking_service.booking.application.dto.request.BookingCreateRequest;
+import com.taken_seat.booking_service.booking.application.service.BookingService;
 import com.taken_seat.common_service.dto.AuthenticatedUser;
 
 @SpringBootTest
